chore: integrate complete decision with optimizations in next sol dpll
