+++ src/main/scala/satify/update/solver/dpll/DpllOneSol.scala
@@ -5,6 +5,7 @@ import satify.model.{CNF, Result, Variable}
 import satify.model.Result.*
 import satify.model.dpll.{Constraint, Decision, DecisionTree}
 import satify.model.dpll.DecisionTree.{Branch, Leaf}
+import satify.update.solver.dpll.DpllDecision.decide
 import satify.update.solver.dpll.DpllOneSol.{dpll, resume}
 import satify.update.solver.dpll.cnf.CNFSat.{isSat, isUnsat}
 import satify.update.solver.dpll.cnf.CNFSimplification.simplifyCnf
@@ -21,14 +22,14 @@ object DpllOneSol:
 
   def dpll(d: Decision): (DecisionTree, Result) =
     if !isUnsat(d.cnf) && !isSat(d.cnf) then
-      randomDecisions(d) match
-        case Some((dl, dr)) =>
-          dpll(dl) match
-            case (dtl, SAT) => (Branch(d, dtl, Leaf(dr)), SAT)
+      decide(d) match
+        case ::(head, next) =>
+          dpll(head) match
+            case (dtl, SAT) => (Branch(d, dtl, Leaf(next.head)), SAT)
             case (dtl, UNSAT) =>
-              dpll(dr) match
+              dpll(next.head) match
                 case (dtr, res) => (Branch(d, dtl, dtr), res)
-        case None => (Leaf(d), if isSat(d.cnf) then SAT else UNSAT)
+        case Nil => (Leaf(d), if isSat(d.cnf) then SAT else UNSAT)
     else (Leaf(d), if isSat(d.cnf) then SAT else UNSAT)
 
   def resume(dt: DecisionTree): (DecisionTree, Result) = dt match
@@ -41,22 +42,7 @@ object DpllOneSol:
     case Branch(d, left, right) =>
       resume(left) match
         case (ldt, SAT) if ldt != left => (Branch(d, ldt, right), SAT)
-        // case (ldt, SAT) if ldt == left =>
         case (ldt, lres) =>
           resume(right) match
             case (rdt, rres) => (Branch(d, ldt, rdt), if lres == SAT then SAT else rres)
-
-  private def randomDecisions(d: Decision): Option[(Decision, Decision)] = d match
-    case Decision(pm, cnf) =>
-      val fv = filterUnconstrVars(pm)
-      if fv.nonEmpty then
-        val v = rnd.nextBoolean()
-        fv(rnd.between(0, fv.size)) match
-          case Variable(n, _) =>
-            Some(
-              (
-                Decision(updateParModel(pm, Constraint(n, v)), simplifyCnf(cnf, Constraint(n, v))),
-                Decision(updateParModel(pm, Constraint(n, !v)), simplifyCnf(cnf, Constraint(n, !v)))
-              )
-            )
-      else None
+
